move eql edit component in a shared folder
