+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation/components/eql_query_edit/eql_query_validator_factory.ts
@@ -9,7 +9,7 @@ import { isEmpty } from 'lodash';
 import type { FormData, ValidationError, ValidationFunc } from '../../../../shared_imports';
 import { KibanaServices } from '../../../../common/lib/kibana';
 import type { EqlOptions } from '../../../../../common/search_strategy';
-import type { FieldValueQueryBar } from '../query_bar';
+import type { FieldValueQueryBar } from '../../../rule_creation_ui/components/query_bar';
 import type { EqlResponseError } from '../../../../common/hooks/eql/api';
 import { EQL_ERROR_CODES, validateEql } from '../../../../common/hooks/eql/api';
 import { EQL_VALIDATION_REQUEST_ERROR } from './translations';
