+++ x-pack/plugins/security_solution/public/detection_engine/rule_creation/components/eql_query_edit/footer.tsx
