+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_edit/fields/eql_query/eql_query_edit_adapter.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import type { DataViewBase } from '@kbn/es-query';
-import { EqlQueryEdit } from '../../../../../../../rule_creation_ui/components/eql_query_edit';
+import { EqlQueryEdit } from '../../../../../../../rule_creation/components/eql_query_edit';
 import type { RuleFieldEditComponentProps } from '../rule_field_edit_component_props';
 import { useDiffableRuleDataView } from '../hooks/use_diffable_rule_data_view';
